As we left it
