+++ CardGame/main.cpp
@@ -9,7 +9,6 @@
 #include <iostream>
 #include <stdlib.h>
 #include <string>
-#include "HandComparator.h"
 
 using namespace std;
 
@@ -61,13 +60,276 @@ int * deck::draw(int playerNo){
 	foo[1] = card;
 	return foo;
 }
+int x;
+class Comparator{
+private:
+	int checkNumPair(int hand[2][5]);
+	int check3Kind(int hand[2][5]);
+	int check4Kind(int hand[2][5]);
+	int checkStraight(int hand[2][5]);
+	int checkFlush(int hand[2][5]);
+	int checkFullHouse(int hand[2][5]);
+	int checkStraightFlush(int hand[2][5]);
+	int checkRoyalFlush(int hand[2][5]);
+public:
+	int getScore(int hand[2][5]);
+	string getHand(int score);
+};
+
+int Comparator::checkNumPair(int hand[2][5])
+{
+	int score = 0;
+	int firstPair = 0;
+	//Check for first pair
+	for (int i = 0; i < 5; i++) {
+		for (int j = i + 1; j < 5; j++) {
+			if (hand[1][i] == hand[1][j]) {
+				firstPair = hand[1][i];
+				score = 1;
+			}
+		}
+	}
+	
+	//Check for second pair.
+	for (int i = 0; i < 5; i++) {
+		for (int j = i + 1; j < 5; j++) {
+			if (hand[1][i] == hand[1][j] && hand[1][i] != firstPair) {
+				score = 2;
+			}
+		}
+	}
+	
+	return score;
+}
+
+int Comparator::check3Kind(int hand[2][5])
+{
+	int score = 0;
+	int counter = 0;
+	
+	for (int i = 0; i < 5; i++) {
+		int value = hand[1][i];
+		for (int i = 0; i < 5; i++) {
+			if (hand[1][i] == value) {
+				counter++;
+				//If three of a kind
+				if (counter == 3) {
+					score = 3;
+					break;
+				}
+			}
+		}
+		
+	}
+	
+	return score;
+}
+
+int Comparator::check4Kind(int hand[2][5])
+{
+	int score = 0;
+	int counter = 0;
+	
+	for (int i = 0; i < 5; i++) {
+		int value = hand[1][i];
+		for (int i = 0; i < 5; i++) {
+			if (hand[1][i] == value) {
+				counter++;
+				//If four of a kind
+				if (counter == 4) {
+					//score = 7;
+					break;
+				}
+			}
+		}
+	}
+	
+	return score;
+}
+
+int Comparator::checkStraight(int hand[2][5])
+{
+	int score = 0;
+	int handVals[5];
+	for (int i = 0; i < 4; i++) {
+		handVals[i] = hand[1][i];
+	}
+	
+	sort(handVals, handVals + 5);
+	
+	if (hand[1][1] == hand[1][0] + 1 && hand[1][2] == hand[1][0] + 2 && hand[1][3] == hand[1][0] + 3 && hand[1][4] == hand[1][0] + 4) {
+		score = 5;
+	}
+	
+	return score;
+}
+
+int Comparator::checkFlush(int hand[2][5])
+{
+	int score = 0;
+	int counter = 0;
+	
+	for (int i = 0; i < 5; i++) {
+		int value = hand[0][i];
+		for (int i = 0; i < 5; i++) {
+			if (hand[0][i] == value) {
+				counter++;
+			}
+			
+			if (counter == 5) {
+				score = 4;
+			}
+		}
+		counter = 0;
+	}
+	
+	
+	
+	return score;
+}
+
+int Comparator::checkFullHouse(int hand[2][5])
+{
+	int score = 0;
+	int counter = 0;
+	int kind = 0;
+	bool checkForPair = false;
+	
+	for (int i = 0; i < 5; i++) {
+		int value = hand[1][i];
+		for (int i = 0; i < 5; i++) {
+			if (hand[1][i] == value) {
+				counter++;
+			}
+			//If three of a kind
+			if (counter == 3) {
+				kind = hand[1][i];
+				checkForPair = true;
+				break;
+			}
+		}
+	}
+	
+	if (checkForPair) {
+		for (int i = 0; i < 5; i++) {
+			for (int j = i + 1; j < 5; j++) {
+				if (hand[1][i] == hand[1][j] && hand[1][i] != kind) {
+					score = 6;
+				}
+			}
+		}
+	}
+	
+	return score;
+}
+
+int Comparator::checkStraightFlush(int hand[2][5])
+{
+	int score = 0;
+	int checkStrt = checkStraight(hand);
+	int checkFlsh = checkFlush(hand);
+	
+	if (checkStrt > 0 && checkFlsh > 0) {
+		score = 8;
+	}
+	return score;
+}
+
+int Comparator::checkRoyalFlush(int hand[2][5]){
+	int score = 0;
+	
+	int handVals[5];
+	for (int i = 0; i < 4; i++) {
+		handVals[i] = hand[1][i];
+	}
+	
+	sort(handVals, handVals + 5);
+	
+	if (hand[1][1] == hand[1][0] + 1 && hand[1][2] == hand[1][0] + 2 && hand[1][3] == hand[1][0] + hand[1][4] == hand[1][0] + 4 && hand[1][1] == 10) {
+		if (checkFlush(hand) > 0) {
+			score = 9;
+		}
+	}
+	
+	return score;
+}
+
+int Comparator::getScore(int hand[2][5]){
+	int score = 0;
+	
+	score = checkNumPair(hand);
+	if (check3Kind(hand) > score) {
+		score = check3Kind(hand);
+	}
+	if (check4Kind(hand) > score) {
+		score = check4Kind(hand);
+	}
+	if (checkStraight(hand) > score) {
+		score = checkStraight(hand);
+	}
+	if (checkFlush(hand) > score) {
+		score = checkFlush(hand);
+	}
+	if (checkFullHouse(hand) > score) {
+		score = checkFullHouse(hand);
+	}
+	if (checkStraightFlush(hand) > score) {
+		score = checkStraightFlush(hand);
+	}
+	if (checkRoyalFlush(hand) > score) {
+		score = checkRoyalFlush(hand);
+	}
+	
+	return score;
+}
+
+string Comparator::getHand(int score){
+	switch (score) {
+		case 0:
+			return "No Pairs";
+			break;
+		case 1:
+			return "One Pair";
+			break;
+		case 2:
+			return "Two Pairs";
+			break;
+		case 3:
+			return "Three of a Kind";
+			break;
+		case 4:
+			return "Straight";
+			break;
+		case 5:
+			return "Flush";
+			break;
+		case 6:
+			return "Full House";
+			break;
+		case 7:
+			return "Four of a Kind";
+			break;
+		case 8:
+			return "Straight Flush";
+			break;
+		case 9:
+			return "Royal Flush";
+			break;
+			
+			
+		default:
+			return "Something done broke!";
+			break;
+	}
+}
 
 int main(void) {
-	char gameState;
+	char gameState = '\0';
 	int numPlayers = 0;
 	string playerNames[5];
 	deck myDeck;
 	
+	while (gameState != 'q'){
 	cout << "Do you wish to play poker or quit (p or q)?" << endl;
 	cin >> gameState;
 	
@@ -103,4 +365,7 @@ int main(void) {
 	Comparator gameComparator;
 	
 	cout << endl << "#####" << endl << gameComparator.getScore(playerHands[0]) << endl << "####";
+		
+	}
+
 }
